Permissions on Users and Decks

+++ src/routes/decks.ts
@@ -9,7 +9,8 @@ import express from "express";
 import { NextFunction, Request, Response } from "express";
 import logger from "morgan";
 import prisma from "../prisma-instance";
-import { deckToApiFormat, getIdFromUrl } from "./decksUtils";
+import { deckToApiFormat, getIdFromUrl } from "../Utils/decksUtils";
+import { fetchUser, isUserLogged } from "../Utils/userUtils";
 import {
 	CardApi,
 	DeckApi,
@@ -18,7 +19,13 @@ import {
 	DeckApi_Createable,
 	CardApi_Createable,
 } from "../type";
-import { bodyValidator } from "./bodyValidator";
+import { bodyValidator } from "../Services/bodyValidator";
+import {
+	getDeckPermissions,
+	getRole,
+	hasPermission,
+	permissionMap,
+} from "../Services/Permissions";
 
 const routerDecks = express.Router();
 const pathDecks = "/decks";
@@ -26,8 +33,20 @@ const pathDecks = "/decks";
 routerDecks.get(
 	"/",
 	async function (req: Request, res: Response, next: NextFunction) {
+		if (!hasPermission(req, "READ_DECKS")) {
+			return res
+				.status(404)
+				.send({ message: "No permission to 'READ_DECKS" });
+		}
 		try {
-			const allDecks: DeckApi_WithoutCards[] = await prisma.deck.findMany();
+			const allDecks: DeckApi_WithoutCards[] = (
+				await prisma.deck.findMany()
+			).map((deck) => {
+				return {
+					...deck,
+					permissions: getDeckPermissions(deck, req.session?.userId),
+				};
+			});
 			res.send(allDecks);
 		} catch (error) {
 			console.error("Error when fetching allDecks: ", error);
@@ -40,6 +59,11 @@ routerDecks.get(
 routerDecks.get(
 	"/:deckId",
 	async function (req: Request, res: Response, next: NextFunction) {
+		if (!hasPermission(req, "READ_DECKS")) {
+			return res
+				.status(404)
+				.send({ message: "No permission to 'READ_DECKS" });
+		}
 		const deckId = getIdFromUrl("deckId", req, res, next);
 
 		try {
@@ -66,7 +90,7 @@ routerDecks.get(
 			if (!deck) {
 				next("Deck not found.");
 			} else {
-				res.send(deckToApiFormat(deck));
+				res.send(deckToApiFormat(deck, req.session?.userId));
 			}
 		} catch (error) {
 			next(error);
@@ -78,6 +102,11 @@ routerDecks.post(
 	"/",
 	bodyValidator("DeckApi_Createable"),
 	async function (req: Request, res: Response, next: NextFunction) {
+		if (!hasPermission(req, "CREATE_DECKS")) {
+			return res
+				.status(404)
+				.send({ message: "No permission to 'CREATE_DECKS" });
+		}
 		const body: DeckApi_Createable = req.body;
 		try {
 			const deck = await prisma.deck.create({
@@ -96,6 +125,11 @@ routerDecks.post(
 							};
 						}),
 					},
+					user: {
+						connect: {
+							id: req.session?.userId,
+						},
+					},
 				},
 				include: {
 					cards: {
@@ -111,9 +145,10 @@ routerDecks.post(
 							deckOrder: "asc",
 						},
 					},
+					user: true,
 				},
 			});
-			res.send(deckToApiFormat(deck));
+			res.send(deckToApiFormat(deck, req.session?.userId));
 		} catch (error) {
 			next(error);
 		}
@@ -123,8 +158,35 @@ routerDecks.put(
 	"/:deckId",
 	bodyValidator("DeckApi"),
 	async function (req: Request, res: Response, next: NextFunction) {
+		if (!hasPermission(req, "EDIT_OWN_DECKS")) {
+			return res
+				.status(404)
+				.send({ message: "No permission to 'EDIT_OWN_DECKS" });
+		}
+		if (!req.session) {
+			throw Error("Session Error");
+		}
+		const userLogged = await fetchUser(req.session.userId);
 		const body: DeckApi = req.body;
 		const deckId = getIdFromUrl("deckId", req, res, next);
+		const deck = await prisma.deck.findUnique({
+			where: {
+				id: deckId,
+			},
+			include: {
+				user: true,
+			},
+		});
+		if (!deck) {
+			return res.status(404).send({
+				message: `Deck ${deckId} not Found.`,
+			});
+		}
+		if (deck.userId != userLogged.id) {
+			return res.status(404).send({
+				message: `Cannot update a deck of another User.`,
+			});
+		}
 		const deleteOp = prisma.card.deleteMany({
 			where: {
 				deckId: {
@@ -177,7 +239,7 @@ routerDecks.put(
 			] = await prisma.$transaction([deleteOp, updateOp]);
 			console.log("gboDebug:[deleteOp_result]", deleteOp_result);
 			console.log("gboDebug:[updateOp_result]", updateOp_result);
-			res.send(deckToApiFormat(updateOp_result));
+			res.send(deckToApiFormat(updateOp_result, req.session?.userId));
 		} catch (error) {
 			console.log("gboDebug:[error]", error);
 			next(error);
@@ -188,6 +250,11 @@ routerDecks.put(
 routerDecks.delete(
 	"/:deckId",
 	async function (req: Request, res: Response, next: NextFunction) {
+		if (!hasPermission(req, "DELETE_OWN_DECKS")) {
+			return res
+				.status(404)
+				.send({ message: "No permission to 'DELETE_OWN_DECKS" });
+		}
 		const deckId = getIdFromUrl("deckId", req, res, next);
 		const deleteCardsOp = prisma.card.deleteMany({
 			where: {
